edits for FS integration
