+++ BasicPpu/fs.html
@@ -1,0 +1,53 @@
+<div class="bg-success py-1 mt-auto">
+        <div class="container px-2">
+            <div class="
+            row
+            align-items-center
+            justify-content-center
+            justify-content-xl-space-around
+            flex-column flex-sm-row
+          ">
+                <div class="col-auto">
+                    <div class="small m-0 text-white text-center">
+                        <strong>Email </strong>
+                        <a class="link-light small"
+                            href="https://office.com">helpdesk@pointpark.edu</a>
+                        <hr>
+                       
+                        <strong>Call </strong>
+                        <a class="link-light small" href="https://office.com">412-392-3494</a>
+                        <hr />
+                        Copyright Point Park University 2022
+                        <hr>
+                        <a class="link-light small" href="https://github.com/ppuhd">PPU Help Desk
+                            Web Dev
+                        </a>
+                    </div>
+                </div>
+                <div class="col-auto">
+                    <div class="small m-0 text-white text-center">
+                        <a class="btn btn-yellow btn-lg px-4 me-sm-3 text-center"
+                            href="index.html"><strong>Submit A
+                                Ticket</strong> </a>
+                    </div>
+                </div>
+
+                <div class="col-auto">
+                    <h2 class="link-light">
+                        <strong>Hours</strong>
+                    </h2>
+                    <h2 class="link-light small">
+                        M-Th 8:30am - 10:30pm
+                    </h2>
+                    <h3 class="link-light small">
+                        F 8:30am - 4:30pm
+                    </h3>
+                    <h2 class="link-light small">
+                        Sat 8:00am - 6:000pm
+                    </h2>
+                    <h2 class=" link-light small">
+                        Sun 10:00am - 6:000pm
+                    </h2>
+                </div>
+            </div>
+        </div>
